% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.R
\name{busparse_gene_count}
\alias{busparse_gene_count}
\title{Get gene count matrix in one step}
\usage{
busparse_gene_count(species, kallisto_out_path, bus_fn, est_ncells,
  est_ngenes, whitelist = NULL, ensembl_version = NULL,
  other_attrs = NULL, save_tr2g = FALSE,
  file_save = "./tr2g_sorted.csv", verbose = TRUE, ncores = 1,
  progress_unit = 5e+06, ...)
}
\arguments{
\item{species}{A character vector of Latin names of species present in this
scRNA-seq dataset. This is used to retrieve Ensembl information from biomart.}

\item{kallisto_out_path}{Path to the \code{kallisto bus} output directory.}

\item{bus_fn}{File name of output of \code{bustools}, namely \code{output.bus}
from \code{kallisto bus} after sorted and converted to text by
\code{bustools}.}

\item{est_ncells}{Estimated number of cells; providing this argument will
speed up computation as it minimizes memory reallocation as vectors grow.}

\item{est_ngenes}{Estimated number of genes.}

\item{whitelist}{A character vector with valid cell barcodes. This is an
optional argument, that defaults to \code{NULL}. When it is \code{NULL},
all cell barcodes present will be included in the sparse matrix, not only
the barcodes known to be valid.}

\item{ensembl_version}{Integer version number of Ensembl (e.g. 94 for the
October 2018 release). This argument defaults to \code{NULL}, which will use
the current release of Ensembl. Use
\code{\link[biomaRt]{listEnsemblArchives}} to see the version number corresponding
to the Ensembl release of a particular date. The version specified here must
match the version of Ensembl where the transcriptome used to build the
kallisto index was downloaded.}

\item{other_attrs}{Character vector. Other attributes to get from Ensembl,
such as gene symbol and position on the genome.}

\item{save_tr2g}{Logical, whether to save the data frame that maps transcripts
to genes to disk.}

\item{file_save}{File name of the file to be saved. If the directory in which
the file is to be saved does not exist, then the directory will be created.}

\item{verbose}{Whether to display progress.}

\item{ncores}{Number of cores to use, defaults to 1.}

\item{progress_unit}{How many iteration to print one progress update when
reading in the \code{kallisto bus} file.}

\item{...}{Othe arguments to be passed to \code{\link[biomaRt]{useEnsembl}},
such as host and mirror.
Use \code{\link[biomaRt]{listAttributes}} to see which attributes are available.}
}
\value{
A sparse matrix with genes in rows and cells in columns.
}
\description{
The \code{bustools} output has 4 columns: barcode, UMI, equivalence class,
and counts. This function directly converts that file into a sparse matrix
that can be used in downstream analyses in one step for species that are in
the Ensembl database. This function condenses a multi-step workflow
implemented in this package into one step. For non-model organisms absent
from Ensembl, please run the individual steps separately, as this function
queries Ensembl through biomart for transcript and gene information required
to aggregate read counts mapped to transcripts into counts for genes. The
vignette has a tutorial of running the individual steps in this workflow.
}
\details{
For 10x data sets, you can find a barcode whitelist file that comes with
CellRanger installation. You don't need to run CellRanger to get that. An
example path to get the whitelist file is
\code{cellranger-2.1.0/cellranger-cs/2.1.0/lib/python/cellranger/barcodes/737K-august-2016.txt}
for v2 chemistry.
}
\note{
By default, this function does not save the data frame that maps
transcripts to genes. As this information, along with human readable gene
names corresponding to gene IDs and other information passed to \code{other_attrs}
that can be retrieved from Ensembl might be useful later in the analysis,
this function has an option to save the data frame. Set \code{save_tr2g = TRUE}
to save this to disk.
}
\concept{functions to generate sparse matrix in one step}
