% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.R
\name{busparse_matrix}
\alias{busparse_matrix}
\title{Get gene count or TCC matrix in one step}
\usage{
busparse_matrix(species, fasta_file, bus_path, est_ncells, est_ngenes,
  whitelist = NULL, gene_count = TRUE, TCC = TRUE,
  ensembl_version = NULL, other_attrs = NULL, save_tr2g = FALSE,
  file_save = "./tr2g_sorted.csv", verbose = TRUE, ncores = 0,
  progress_unit = 5e+06, ...)
}
\arguments{
\item{species}{A character vector of Latin names of species present in this
scRNA-seq dataset. This is used to retrieve Ensembl information from biomart.}

\item{fasta_file}{Character vector of paths to the transcriptome FASTA files
used to build the kallisto index. Exactly one of \code{species} and
\code{fasta_file} can be missing.}

\item{bus_path}{Path to the sorted text \code{bus} output file.}

\item{est_ncells}{Estimated number of cells; providing this argument will
speed up computation as it minimizes memory reallocation as vectors grow.}

\item{est_ngenes}{Estimated number of genes or equivalence classes.}

\item{whitelist}{A character vector with valid cell barcodes. This is an
optional argument, that defaults to \code{NULL}. When it is \code{NULL},
all cell barcodes present will be included in the sparse matrix whether they
are known to be valid or not.}

\item{gene_count}{Logical, whether the gene count matrix should be returned.}

\item{TCC}{Logical, whether the TCC matrix should be returned.}

\item{ensembl_version}{Integer version number of Ensembl (e.g. 94 for the
October 2018 release). This argument defaults to \code{NULL}, which will use
the current release of Ensembl. Use
\code{\link[biomaRt]{listEnsemblArchives}} to see the version number corresponding
to the Ensembl release of a particular date. The version specified here must
match the version of Ensembl where the transcriptome used to build the
kallisto index was downloaded.}

\item{other_attrs}{Character vector. Other attributes to get from Ensembl,
such as gene symbol and position on the genome.}

\item{save_tr2g}{Logical, whether to save the data frame that maps transcripts
to genes to disk.}

\item{file_save}{File name of the file to be saved. If the directory in which
the file is to be saved does not exist, then the directory will be created.}

\item{verbose}{Whether to display progress.}

\item{ncores}{Number of cores to use, defaults to 0, which means the system
will automatically determine the number of cores as it sees fit. Negative
numbers are interpreted as 0. Positive numbers will limit the number of cores
used.}

\item{progress_unit}{How many iteration to print one progress update when
reading in the \code{kallisto bus} file.}

\item{...}{Othe arguments to be passed to \code{\link[biomaRt]{useEnsembl}},
such as host and mirror.
Use \code{\link[biomaRt]{listAttributes}} to see which attributes are available.}
}
\value{
If both gene count and TCC matrices are returned, then this function
returns a list with two matrices, each with genes/equivalence classes in the
rows and barcodes in the columns. If only one of gene count and TCC matrices
is returned, then a \code{dgCMatrix} with genes/equivalence classes in the
rows and barcodes in the columns.
}
\description{
The \code{bustools} output has 4 columns: barcode, UMI, equivalence class,
and counts. This function directly converts that file into a sparse matrix
that can be used in downstream analyses in one step for species that are in
the Ensembl database. This function condenses a multi-step workflow
implemented in this package into one step. For non-model organisms absent
from Ensembl, please run the individual steps separately, as this function
either queries Ensembl through biomart or parses Ensembl FASTA sequence names
for transcript and gene information required to aggregate read counts mapped
to transcripts into counts for genes. The vignette has a tutorial of running
the individual steps in this workflow.
}
\details{
For 10x data sets, you can find a barcode whitelist file that comes with
CellRanger installation. You don't need to run CellRanger to get that. An
example path to get the whitelist file is
\code{cellranger-2.1.0/cellranger-cs/2.1.0/lib/python/cellranger/barcodes/737K-august-2016.txt}
for v2 chemistry.

Passing FASTA files to \code{fasta_file} is faster than passing \code{species}
since with the latter, this function will query the Ensembl biomart database,
which is really slow.

By default, this function does not save the data frame that maps
transcripts to genes. As this information, along with human readable gene
names corresponding to gene IDs and other information passed to \code{other_attrs}
that can be retrieved from Ensembl might be useful later in the analysis,
this function has an option to save the data frame. Set \code{save_tr2g = TRUE}
to save this to disk.
}
\concept{functions to generate sparse matrix in one step}
