\name{BUStoolsR-package}
\alias{BUStoolsR-package}
\alias{BUStoolsR}
\docType{package}
\title{
\packageTitle{BUStoolsR}
}
\description{
\packageDescription{BUStoolsR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BUStoolsR}
\packageIndices{BUStoolsR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{BUStoolsR}

Maintainer: \packageMaintainer{BUStoolsR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
