% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotypes.R
\docType{data}
\name{cellranger_biotypes}
\alias{cellranger_biotypes}
\title{Cell Ranger gene biotypes}
\format{
A character vector with all Cell Ranger reference package gene
biotypes.
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/advanced/references}
}
\usage{
cellranger_biotypes
}
\description{
In the GRCh38 Cell Ranger reference package, an Ensembl human GTF file is
filtered by gene biotypes. This vector includes all gene biotypes included by
this Cell Ranger reference package.
}
\seealso{
ensembl_gene_biotypes ensembl_tx_biotypes
}
\keyword{datasets}
