% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_char1}
\alias{check_char1}
\title{Check that an object is a character vector of length 1}
\usage{
check_char1(x)
}
\arguments{
\item{x}{Named vector of arguments to be checked.}
}
\value{
Error if \code{x} is not a character vector with length 1.
}
\description{
Just in case the user passes something with length more than 1 and messes up
everything thanks to vectorization.
}
