% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file_gz}
\alias{check_file_gz}
\title{Check for existence of file that may be compressed}
\usage{
check_file_gz(fn)
}
\arguments{
\item{fn}{Path to the file.}
}
\value{
The path itself if the file exists, the path with gz extension if the
gzipped version exists, and an error if neither exists.
}
\description{
Functions like \code{fread} can directly read gzipped files, so there's no need to
decompress. This function checks if a file exists, and if it doesn't, whether
its gzipped counterpart exists.
}
