% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_tag_present}
\alias{check_tag_present}
\title{Check that a tag is present in attribute field of GTF/GFF}
\usage{
check_tag_present(tags_use, tags, error = TRUE)
}
\arguments{
\item{tags_use}{The tags to be checked.}

\item{tags}{The tags present in attribute field.}
}
\value{
Error or warning if tag is absent.
}
\description{
The attribute field of GTF/GFF files are very complicated and is very
inconsistent between sources. This function is to make sure that transcript
and gene IDs can be extracted properly.
}
