% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{check_tx}
\alias{check_tx}
\title{Check if transcript ID in transcriptome and annotation match}
\usage{
check_tx(tx_annot, tx)
}
\arguments{
\item{tx_annot}{Character vector of transcript IDs from the annotation.}

\item{tx}{Character vector of transcript IDs from the transcriptome.}
}
\value{
Character vector of the overlapping transcript IDs.
}
\description{
This function throws an error if transcript IDs in transcriptome and
annotation do not overlap. If they do overlap, this function will give a
message about transcript IDs that do not agree in the transcriptome and the
annotation
}
