% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{collapse_isoforms}
\alias{collapse_isoforms}
\title{Collapse different isoforms of a gene}
\usage{
collapse_isoforms(gr, gene_id = "gene_id",
  transcript_id = "transcript_id", exon_number = "exon_number")
}
\arguments{
\item{gr}{A \code{GRanges} object with only exon ranges. As this function is for
internal use only and the fields are checked elsewhere, fields will not be
checked in this function.}

\item{gene_id}{Name of the metadata field for gene ID.}

\item{transcript_id}{Name of the metadata field for transcript ID.}

\item{exon_number}{Name of the metadata field for exon number.}
}
\value{
A \code{GRanges} object with only exon ranges and different isoforms
collapsed. Now the "transcript ID" in the metadata columns will be the same
as the gene ID.
}
\description{
This function takes the union of all exon ranges of different isoform of each
gene.
}
