% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr2g.R
\name{.tr2g_fasta}
\alias{.tr2g_fasta}
\title{Get transcript and gene info from names in FASTA files}
\usage{
.tr2g_fasta(file, verbose = TRUE)
}
\arguments{
\item{file}{Path to the FASTA file to be read. The file can remain gzipped.}

\item{verbose}{Whether to display progress.}
}
\value{
A data frame with 3 columns: \code{gene} for gene ID, \code{transcript}
for transcript ID, and \code{gene_name} for gene names.
}
\description{
FASTA files, such as those for cDNA and ncRNA from Ensembl, have genome
annotation information in the name of each sequence entry. This function
extracts the transcript and gene IDs from such FASTA files.
}
\details{
At present, this function only works with FASTA files from Ensembl, and uses
regex to extract Ensembl IDs. Sequence names should be formatted as follows:\preformatted{ENST00000632684.1 cdna chromosome:GRCh38:7:142786213:142786224:1 
gene:ENSG00000282431.1 gene_biotype:TR_D_gene transcript_biotype:TR_D_gene 
gene_symbol:TRBD1 description:T cell receptor beta diversity 1 
[Source:HGNC Symbol;Acc:HGNC:12158]
}

If your FASTA file sequence names are formatted differently, then you must
extract the transcript and gene IDs by some other means. The Bioconductor
package \code{Biostrings} is recommended; after reading the FASTA file into
R, the sequence names can be accessed by the \code{names} function.
}
\seealso{
\code{\link{tr2g_fasta}}
}
