% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr2g.R
\name{EC2gene}
\alias{EC2gene}
\title{Map EC Index to Genes Compatible with the EC}
\usage{
EC2gene(tr2g, kallisto_out_path, ncores = 0, verbose = TRUE)
}
\arguments{
\item{tr2g}{A Data frame with columns \code{gene} and \code{transcript}, in
the same order as in the transcriptome index for \code{kallisto}.}

\item{kallisto_out_path}{Path to the \code{kallisto bus} output directory.}

\item{ncores}{Number of cores to use, defaults to 0, which means the system
will automatically determine the number of cores as it sees fit. Negative
numbers are interpreted as 0. Positive numbers will limit the number of cores
used.}

\item{verbose}{Whether to display progress. Defaults to \code{TRUE}.}
}
\value{
A named list each element of whom is the set of genes the
corresponding EC is compatible to. The names are the EC indices, which are
also row names in the TCC matrix.
}
\description{
In the output file \code{output.bus}, equivalence classes (EC) are denoted by
an index, which is related to the set of transcripts the EC is compatible to
in the output file \code{matrix.ec}. This function further relates the set of
transcripts to the set of genes the EC is compatible to. This function first
reads in \code{matrix.ec}, and then translates the transcripts into genes.
}
\details{
The data frame passed to \code{tr2g} can be generated from function
\code{\link{transcript2gene}} in this package for any organism that has gene and
transcript ID on Ensembl, or from the \code{tr2g_*} family of function.
You no longer need to use this function before running \code{make_sparse_matrix};
the purpose of this function is to query which genes equivalence classes map
to.

This function returns a named list, whose names are the equivalence class (EC)
indices in the first column of file \code{matrix.ec} in the kallisto bus output.
Note that the EC indices are 0 based, so are the names of elements of this
list. The names of elements of this list are 0 based EC indices as string.
}
\examples{
\dontrun{
# Download dataset already in BUS format
library(TENxhgmmBUS)
download_hgmm(".", "hgmm100")
tr2g <- transcript2gene(c("Homo sapiens", "Mus musculus"),
                        kallisto_out_path = "./out_hgmm100")
genes <- EC2gene(tr2g, "./out_hgmm100", ncores = 1, verbose = FALSE)
}

}
\seealso{
\code{\link{transcript2gene}}
}
