% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotypes.R
\docType{data}
\name{ensembl_gene_biotypes}
\alias{ensembl_gene_biotypes}
\title{Gene biotypes from Ensembl}
\format{
A character vector with all Ensembl gene biotypes.
}
\source{
This vector is all the unique gene biotypes in the Ensembl version 99
human GTF file.
}
\usage{
ensembl_gene_biotypes
}
\description{
These vectors are here to make it easy to look up which biotypes are
available for Ensembl without having to parse GTF and fasta files every time.
}
\keyword{datasets}
