% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotypes.R
\docType{data}
\name{ensembl_gff_mcols}
\alias{ensembl_gff_mcols}
\title{These are the column names of the \code{mcols} when the Ensembl GTF file is read
into R as a \code{GRanges}, including \code{gene_id}, \code{transcript_id}, \code{biotype},
\code{description}, and so on, and the mandatory tags like \code{ID}, \code{Name}, and
\code{Parent}.}
\format{
A character vector
}
\source{
Ensembl version 99 human GFF3 file
}
\usage{
ensembl_gff_mcols
}
\description{
These are the column names of the \code{mcols} when the Ensembl GTF file is read
into R as a \code{GRanges}, including \code{gene_id}, \code{transcript_id}, \code{biotype},
\code{description}, and so on, and the mandatory tags like \code{ID}, \code{Name}, and
\code{Parent}.
}
\keyword{datasets}
