% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotypes.R
\docType{data}
\name{ensembl_gtf_mcols}
\alias{ensembl_gtf_mcols}
\title{Tags in the attributes field of Ensembl GTF files}
\format{
A character vector
}
\source{
Ensembl version 99 human GTF file
}
\usage{
ensembl_gtf_mcols
}
\description{
These are the column names of the \code{mcols} when the Ensembl GTF file is read
into R as a \code{GRanges}, including \code{gene_id}, \code{transcript_id}, \code{gene_biotype},
\code{transcript_biotype}, \code{description}, and so on.
}
\keyword{datasets}
