% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{exclude_short_exons}
\alias{exclude_short_exons}
\title{Exclude exons that are too short}
\usage{
exclude_short_exons(gr, L, gene_id = "gene_id",
  transcript_id = "transcript_id", exon_number = "exon_number",
  exon_id = "exon_id")
}
\arguments{
\item{gr}{A \code{GRanges} object with only exon ranges. As this function is for
internal use only and the fields are checked elsewhere, fields will not be
checked in this function.}

\item{L}{Length of the biological read. For instance, 10xv1: 98 nt,
10xv2: 98 nt, 10xv3: 91 nt, Drop-seq: 50 nt. If in doubt check read length
in a fastq file for biological reads with the \code{bash} commands:
If the fastq file is gzipped, then do \code{zcat your_file.fastq.gz | head} on
Linux. If on Mac, then \code{zcat < your_file.fastq.gz | head}. Then you will see
lines with nucleotide bases. Copy one of those lines and determine its length
with \code{\link{str_length}} in R or \code{echo -n <the sequence> | wc -c} in
\code{bash}. Which file corresponds to biological reads depends on the particular
technology.}

\item{gene_id}{Name of the metadata field for gene ID.}

\item{transcript_id}{Name of the metadata field for transcript ID.}

\item{exon_number}{Name of the metadata field for exon number.}

\item{exon_id}{Character vector of length 1. Tag in \code{attribute}
field corresponding to exon ID. If the specified tag does not exist in the
GTF file, then a new metadata column that can uniquely identify exons will
be created.}
}
\value{
A \code{GRanges} object without exons that are too short.
}
\description{
Exons that are too short are excluded from the \code{GRanges} object, except the
first and the last exons of a transcript.
}
