% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{extract_tx}
\alias{extract_tx}
\title{Extract transcriptomic sequences from genome}
\usage{
extract_tx(Genome, annot)
}
\arguments{
\item{Genome}{Either a \code{\link{BSgenome}} or a \code{\link{XStringSet}}
object of genomic sequences, where the intronic sequences will be extracted
from. Use \code{\link{genomeStyles}} to check which styles are supported for
your organism of interest; supported styles can be interconverted. If the
style in your genome or annotation is not supported, then the style of
chromosome names in the genome and annotation should be manually set to be
consistent.}

\item{annot}{A \code{GRanges} object.}
}
\value{
A \code{DNAStringSet} object with transcript sequences, whose names are
the transcript IDs.
}
\description{
This is a thin wrapper around \code{\link{extractTranscriptSeqs}}. This
wrapper takes care of object type of the genome.
}
