% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_args}
\alias{get_args}
\title{Get values passed to arguments in a function}
\usage{
get_args()
}
\value{
A named list mapping arguments to values passed to them.
}
\description{
This function is for giving informative error messages about which argument
is wrong. This is from the StackOverflow question
\url{https://stackoverflow.com/questions/17256834/getting-the-arguments-of-a-parent-function-in-r-with-names}
}
