% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{get_flank_lengths}
\alias{get_flank_lengths}
\title{Get lengths of flanking regions}
\usage{
get_flank_lengths(df, L)
}
\arguments{
\item{df}{A data frame with the following columns:
\describe{
\item{chr}{Which chromosome the gene is from.}
\item{start}{Start of the exonic range. For genes on the minus strand, this
is the 5' end on the plus strand, and thus actually the end of the exonic
range.}
\item{end}{End of the exonic range.}
\item{strand}{Strand of the transcript of interest.}
\item{gene_id}{Gene ID.}
\item{transcript_id}{Transcript ID.}
\item{exon_number}{Order of the exon from 5' to 3' on the transcript. Should
be integer.}
}
The column names must be as described above.}

\item{L}{Length of the biological read. For instance, 10xv1: 98 nt,
10xv2: 98 nt, 10xv3: 91 nt, Drop-seq: 50 nt. If in doubt check read length
in a fastq file for biological reads with the \code{bash} commands:
If the fastq file is gzipped, then do \code{zcat your_file.fastq.gz | head} on
Linux. If on Mac, then \code{zcat < your_file.fastq.gz | head}. Then you will see
lines with nucleotide bases. Copy one of those lines and determine its length
with \code{\link{str_length}} in R or \code{echo -n <the sequence> | wc -c} in
\code{bash}. Which file corresponds to biological reads depends on the particular
technology.}
}
\value{
A data frame with the following columns:
\describe{
\item{gene_id}{Gene ID.}
\item{intron_id}{Intron ID, constructed from transcript ID and the order
of introns from 5' annd 3' of the transcript.}
\item{flank5}{Length of the flanking region on the 5' end of the intron.}
\item{flank3}{Length of the flanking region on the 3' end of the intron.}
}
}
\description{
Get lengths of flanking regions at both 5' and 3' ends of each intron, and
sort the flanking lengths in the order the introns will be sorted.
}
