% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{get_flanked_introns}
\alias{get_flanked_introns}
\title{Get flanked intronic ranges}
\usage{
get_flanked_introns(txdb, metas)
}
\arguments{
\item{txdb}{A \code{TxDb} object with exonic annotations.}

\item{metas}{The data frame returned by \code{\link{get_flank_lengths}}.}
}
\value{
A \code{GRanges} object with ranges for flanked intronic regions.
}
\description{
Get flanked intronic ranges
}
