% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knee_plot.R
\name{get_inflection}
\alias{get_inflection}
\title{Compute inflection point of knee plot}
\usage{
get_inflection(df, lower = 100)
}
\arguments{
\item{df}{The data frame from \code{\link{get_knee_df}}.}

\item{lower}{Minimum total UMI counts for barcode for it to be considered
when calculating the inflection point; this helps to avoid the noisy part of
the curve for barcodes with very few counts.}
}
\value{
A \code{numeric(1)} for the total UMI count at the inflection point.
}
\description{
Compute inflection point of knee plot
}
\note{
Code in part adapted from \code{barcodeRanks} from \code{DropetUtils}.
}
