% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{get_intron_flanks}
\alias{get_intron_flanks}
\title{Get flanked intronic ranges}
\usage{
get_intron_flanks(grl, L, get_junctions)
}
\arguments{
\item{grl}{A \code{CompressedGRangesList} for exonic ranges, each element for one
transcript.}

\item{L}{Read length.}

\item{get_junctions}{Logical, whether to also return exon-exon junctions.}
}
\value{
If \code{get_junctions} is \code{FALSE}, then a \code{GRanges} object
with ranges for flanked intronic regions. If \code{get_junctions} is \code{TRUE}, then
in addition to the flanked intronic ranges, a \code{CompressedGRangesList} with
exon-exon junction ranges and ranges for transcripts without introns.
}
\description{
Adapted from Julien Roux.
}
