% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_cellranger}
\alias{read_cellranger}
\title{Read sparse matrix from CellRanger format}
\usage{
read_cellranger(out_path)
}
\arguments{
\item{out_path}{Directory where \code{matrix.mtx}, \code{features.tsv}, and \code{barcodes.tsv}
are saved.}
}
\value{
A sparse matrix as in \code{matrix.mtx} with contents of \code{features.tsv} as
row names and contents of \code{barcodes.tsv} as column names.
}
\description{
This function reads \code{matrix.mtx}, \code{features.tsv}, and \code{barcodes.tsv}. The
files can be gzipped for compression, and supplied to this function without
decompression.
}
