% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotypes.R
\docType{data}
\name{refseq_gff_mcols}
\alias{refseq_gff_mcols}
\title{Tags in the attributes field of RefSeq GFF files}
\format{
A character vector
}
\source{
Ensembl version 99 human GTF file
}
\usage{
refseq_gff_mcols
}
\description{
These are the column names of the \code{mcols} when the Ensembl GTF file is read
into R as a \code{GRanges}, including \code{gene}, \code{transcript_id}, \code{gene_biotype},
\code{description}, and so on, and the mandatory tags like \code{ID}, \code{Name}, and
\code{Parent}.
}
\keyword{datasets}
