% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_cellranger}
\alias{save_cellranger}
\title{Save sparse matrix in CellRanger \code{mtx} format}
\usage{
save_cellranger(mat, out_path, compress = TRUE, compression_level = 6)
}
\arguments{
\item{mat}{The sparse matrix, as \code{\link[Matrix]{dgCMatrix-class}}.}

\item{out_path}{Path to directory to save the file. If the directory does not
exist, then it will be created.}

\item{compress}{Logical, whether to compress the output files. If \code{TRUE},
then each file will be compressed individually. The directory containing the
files will not be compressed as a whole.}

\item{compression_level}{Compression level for gzip.}
}
\value{
Nothing is returned into the R session; the files are written on disk.
}
\description{
This function saves the sparse matrix generated by \code{\link{make_sparse_matrix}},
\code{\link{busparse_gene_count}} in the CellRanger output format. The sparse
matrix itself will be saved as \code{matrix.mtx}. Barcodes will be saved as
\code{barcodes.tsv}. Gene names (for gene count matrices) or ECs (for TCC
matrices) will be saved as \code{features.tsv}. The purpose of this function
is interoperability, so coworkers who do not use R can load the output of
this package for downstream analysis.
}
