% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{standardize_tags}
\alias{standardize_tags}
\title{Standardize GRanges field names}
\usage{
standardize_tags(gr, gene_id, transcript_id, exon_number)
}
\arguments{
\item{gr}{A \code{GRanges} object.}

\item{gene_id}{Name of the metadata field for gene ID.}

\item{transcript_id}{Name of the metadata field for transcript ID.}

\item{exon_number}{Name of the metadata field for exon number.}
}
\value{
A \code{GRanges} object with standardized names: gene ID as \code{gene_id},
transcript ID as \code{transcript_id}, and exon number as \code{exon_number}.
}
\description{
To avoid introducing rlang as another dependency for tidyeval. This function
will also convert exon numbers to integer.
}
