% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_methods.R
\docType{data}
\name{styles}
\alias{styles}
\title{List of all possible chromosome naming styles}
\format{A character vector}
\usage{
styles
}
\description{
List of all possible chromosome naming styles
}
\keyword{datasets}
