% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tx.R
\name{sub_annot}
\alias{sub_annot}
\title{Remove chromosomes in anotation absent from genome}
\usage{
sub_annot(chrs_use, annot)
}
\arguments{
\item{chrs_use}{Character vector of names of chromosomes present in both the
annotation and the genome.}

\item{annot}{Either a \code{GRanges} object or a \code{TxDb} object for gene
annotation.}
}
\value{
A subsetted genome annotation of the same type ofo the input genome
annotation.
}
\description{
Remove chromosomes in anotation absent from genome
}
