% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr2g.R
\name{tr2g_ensembl}
\alias{tr2g_ensembl}
\title{Get transcript and gene info from Ensembl}
\usage{
tr2g_ensembl(species, type = "vertebrate", other_attrs = NULL,
  use_transcript_version = TRUE, use_gene_version = TRUE,
  ensembl_version = NULL, verbose = TRUE, ...)
}
\arguments{
\item{species}{Character vector of length 1, Latin name of the species of
interest.}

\item{type}{Character, must be one of "vertebrate", "metazoa", "plant",
"fungus" and "protist". Passing "vertebrate" will use the default
www.ensembl.org host. Gene annotation of some common invertebrate model
organisms, such as \emph{Drosophila melanogaster}, are available on www.ensembl.org
so for these invertebrate model organisms, "vertebrate" can be used for this
argument. Passing values other than "vertebrate" will use other Ensembl hosts.
For animals absent from www.ensembl.org, try "metazoa".}

\item{other_attrs}{Character vector. Other attributes to get from Ensembl,
such as gene symbol and position on the genome.
Use \code{\link[biomaRt]{listAttributes}} to see which attributes are available.}

\item{use_transcript_version}{Logical, whether to include version number in
the Ensembl transcript ID.}

\item{use_gene_version}{Logical, whether to include version number in the
Ensembl gene ID.}

\item{ensembl_version}{Integer version number of Ensembl (e.g. 94 for the
October 2018 release). This argument defaults to \code{NULL}, which will use
the current release of Ensembl. Use
\code{\link[biomaRt]{listEnsemblArchives}} to see the version number corresponding
to the Ensembl release of a particular date. The version specified here must
match the version of Ensembl where the transcriptome used to build the
kallisto index was downloaded.}

\item{verbose}{Whether to display progress.}

\item{\dots}{Othe arguments to be passed to \code{\link[biomaRt]{useEnsembl}},
such as host and mirror.}
}
\value{
A data frame with 3 columns: \code{gene} for gene ID, \code{transcript}
for transcript ID, and \code{gene_name} for gene names. If \code{other_attrs}
has been specified, then those will also be columns in the data frame returned.
}
\description{
This function queries Ensembl biomart to convert transcript IDs to gene IDs.
}
\examples{
tr2g <- tr2g_ensembl(species = "Felis catus", other_attrs = "description")
tr2g <- tr2g_ensembl(species = "Arabidopsis thaliana", type = "plant")

}
\seealso{
Other functions to retrieve transcript and gene info: \code{\link{sort_tr2g}},
  \code{\link{tr2g_fasta}}, \code{\link{tr2g_gff3}},
  \code{\link{tr2g_gtf}}, \code{\link{transcript2gene}}
}
\concept{functions to retrieve transcript and gene info}
