% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{tr2g_junction}
\alias{tr2g_junction}
\title{tr2g for exon-exon junctions}
\usage{
tr2g_junction(tr2g_cdna, junction_names)
}
\arguments{
\item{tr2g_cdna}{The original \code{tr2g_cdna}.}

\item{junction_names}{Names of junctions internally generated.}
}
\value{
A \code{tr2g} data frame where "transcripts" are the exon-exon junctions
and genes are the corresponding genes.
}
\description{
tr2g for exon-exon junctions
}
