% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{write_velocity_output}
\alias{write_velocity_output}
\title{Write the files for RNA velocity to disk}
\usage{
write_velocity_output(out_path, introns, Genome, Transcriptome,
  isoform_action, tr2g_cdna, compress_fa, width)
}
\arguments{
\item{out_path}{Directory to save the outputs written to disk. If this
directory does not exist, then it will be created.}

\item{introns}{Intronic ranges plus flanking region, returned by
\code{\link{get_flanked_introns}}.}

\item{tr2g_cdna}{A data frame with columns \code{transcript} and \code{gene} that maps
transcripts to genes for spliced transcripts.}
}
\value{
Nothing into the R session. The files are written to disk.
}
\description{
Write the files for RNA velocity to disk, in the specified output directory.
}
