% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tx.R
\name{check_genome}
\alias{check_genome}
\title{Check for chromosomes in genome but not annotation}
\usage{
check_genome(chrs_use, genome)
}
\arguments{
\item{chrs_use}{Character vector of names of chromosomes present in both the
annotation and the genome.}

\item{genome}{Either a \code{\link{BSgenome}} or a \code{\link{XStringSet}}
object of genomic sequences, where the intronic sequences will be extracted
from. Naming convention of chromosomes in the genome must match that in the
annotation, e.g. consistently use 1 for chromosome 1 or consistently use chr1.}
}
\value{
Nothing. Will emit message if the genome contains chromosomes absent
from the annotation.
}
\description{
Check for chromosomes in genome but not annotation
}
