% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr2g.R
\name{EC2gene}
\alias{EC2gene}
\title{Map EC Index to Genes Compatible with the EC}
\usage{
EC2gene(tr2g, kallisto_out_path, ncores = 1, verbose = TRUE)
}
\arguments{
\item{tr2g}{A Data frame with columns \code{gene} and \code{transcript}, in
the same order as in the transcriptome index for \code{kallisto}.}

\item{kallisto_out_path}{Path to the \code{kallisto bus} output directory.}

\item{ncores}{Number of cores to use, defaults to 1.}

\item{verbose}{Whether to display progress, though this function usually
does not take very long. Defaults to \code{TRUE}.}
}
\value{
A list each element of whom is the set of genes the corresponding EC
is compatible to. The genes are in Ensembl ID with version number. The 
elements of this list are in the same order as the ECs listed in the
\code{kallisto bus} output file \code{matrix.ec}.
}
\description{
In the output file \code{output.bus}, equivalence classes (EC) are denoted by
an index, which is related to the set of transcripts the EC is compatible to 
in the output file \code{matrix.ec}. This function further relates the set of
transcripts to the set of genes the EC is compatible to. This function first
reads in \code{matrix.ec}, and then translates the transcripts into genes.
}
\details{
The data frame passed to \code{tr2g} can be generated from function 
\code{transcript2gene} in this package for any organism that has gene and
transcript ID on Ensembl. The output of this function should be passed to
\code{make_sparse_matrix} in the next step in the workflow, which will 
produce the sparse matrix that can be used in \code{Seurat}.
}
\seealso{
\code{\link{transcript2gene}}
}
