% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tx.R
\docType{methods}
\name{extract_tx}
\alias{extract_tx}
\alias{extract_tx,ANY,GRanges-method}
\alias{extract_tx,ANY,TxDb-method}
\title{Extract transcriptomic sequences from genome}
\usage{
extract_tx(genome, annot, ...)

\S4method{extract_tx}{ANY,GRanges}(genome, annot,
  transcript_id = "transcript_id")

\S4method{extract_tx}{ANY,TxDb}(genome, annot)
}
\arguments{
\item{genome}{Either a \code{\link{BSgenome}} or a \code{\link{XStringSet}}
object of genomic sequences, where the intronic sequences will be extracted
from. Naming convention of chromosomes in the genome must match that in the
annotation, e.g. consistently use 1 for chromosome 1 or consistently use chr1.}

\item{annot}{Either a \code{GRanges} object or a \code{TxDb} object for gene
annotation.}

\item{\dots}{Extra argument for methods.}

\item{transcript_id}{Character vector of length 1. Tag in \code{attribute}
field corresponding to transcript IDs. This argument must be supplied and
cannot be \code{NA} or \code{NULL}. Will throw error if tag indicated in this
argument does not exist.}
}
\value{
A \code{DNAStringSet} object with transcript sequences, whose names are
the transcript IDs.
}
\description{
This is a thin wrapper around \code{\link{extractTranscriptSeqs}}. This
wrapper takes care of object type of the genome.
}
