% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{get_flanked_introns}
\alias{get_flanked_introns}
\title{Get flanked intronic ranges}
\usage{
get_flanked_introns(gr, L)
}
\arguments{
\item{gr}{Exon \code{GRangesList}, each element of which is for a transcript or a
gene.}

\item{L}{Read length.}
}
\value{
A \code{GRangesList} object with ranges for flanked intronic regions.
Each element of the list is for a transcript or a gene.
}
\description{
Adapted from Julien Roux.
}
