% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knee_plot.R
\name{knee_plot}
\alias{knee_plot}
\alias{knee_plot,DataFrame-method}
\alias{knee_plot,list-method}
\title{Knee plot for filtering empty droplets}
\usage{
knee_plot(bc_rank)

\S4method{knee_plot}{DataFrame}(bc_rank)

\S4method{knee_plot}{list}(bc_rank)
}
\arguments{
\item{bc_rank}{A \code{DataFrame} output from \code{DropletUtil::barcodeRanks} or a
named list of such \code{DataFrame}s.}
}
\value{
A ggplot2 object.
}
\description{
Visualizes the inflection point to filter empty droplets. This function plots
different datasets with a different color. Facets can be added after calling
this function with \verb{facet_*} functions. Will be added to the next release
version of BUSpaRse.
}
\examples{
library(DropletUtils)
set.seed(2000)
my.counts <- DropletUtils:::simCounts()
br.out <- barcodeRanks(my.counts)
knee_plot(br.out)

}
