% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{species2dataset}
\alias{species2dataset}
\title{Convert Latin species name to dataset name}
\usage{
species2dataset(species)
}
\arguments{
\item{species}{A character vector of Latin names of species present in this
scRNA-seq dataset.}
}
\value{
The appropriate dataset name for biomart.
}
\description{
This function converts Latin species name to a dataset name in biomart to
query gene and transcript ID.
}
