% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{species2dataset}
\alias{species2dataset}
\title{Convert Latin species name to dataset name}
\usage{
species2dataset(
  species,
  type = c("vertebrate", "metazoa", "plant", "fungus", "protist")
)
}
\arguments{
\item{species}{A character vector of Latin names of species present in this
scRNA-seq dataset. This is used to retrieve Ensembl information from biomart.}

\item{type}{A character vector indicating the type of each species. Each
element must be one of "vertebrate", "metazoa", "plant", "fungus", and
"protist". If length is 1, then this type will be used for all species specified
here. Can be missing if \code{fasta_file} is specified.}
}
\value{
The appropriate dataset name for biomart.
}
\description{
This function converts Latin species name to a dataset name in biomart to
query gene and transcript ID.
}
\examples{
species2dataset(species = "Homo sapiens")
}
