% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tx.R
\docType{methods}
\name{subset_annot}
\alias{subset_annot}
\alias{subset_annot,DNAStringSet-method}
\alias{subset_annot,BSgenome-method}
\title{Subset genome annotation}
\usage{
subset_annot(genome, annot)

\S4method{subset_annot}{DNAStringSet}(genome, annot)

\S4method{subset_annot}{BSgenome}(genome, annot)
}
\arguments{
\item{genome}{Either a \code{\link{BSgenome}} or a \code{\link{XStringSet}}
object of genomic sequences, where the intronic sequences will be extracted
from. Naming convention of chromosomes in the genome must match that in the
annotation, e.g. consistently use 1 for chromosome 1 or consistently use chr1.}

\item{annot}{Either a \code{GRanges} object or a \code{TxDb} object for gene
annotation.}
}
\value{
A subsetted genome annotation of the same type ofo the input genome
annotation.
}
\description{
Exclude chromosomes present in the annotation but absent from the genonme.
}
