% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr2g.R
\name{tr2g_EnsDb}
\alias{tr2g_EnsDb}
\title{Get transcript and gene info from EnsDb objects}
\usage{
tr2g_EnsDb(ensdb, other_attrs = NULL, use_gene_name = TRUE,
  use_transcript_version = TRUE, use_gene_version = TRUE)
}
\arguments{
\item{ensdb}{Ann \code{EnsDb} object, such as from \code{AnnotationHub} or
\code{EnsDb.Hsapiens.v86}.}

\item{other_attrs}{Character vector. Other attributes to get from the \code{EnsDb}
object, such as gene symbol and position on the genome.
Use \code{\link{columns}} to see which attributes are available.}

\item{use_gene_name}{Logical, whether to get gene names.}

\item{use_transcript_version}{Logical, whether to include version number in
the Ensembl transcript ID. To decide whether to
include transcript version number, check whether version numbers are included
in the \code{transcripts.txt} in the \code{kallisto} output directory. If that file
includes version numbers, then trannscript version numbers must be included
here as well. If that file does not include version numbers, then transcript
version numbers must not be included here.}

\item{use_gene_version}{Logical, whether to include version number in the
Ensembl gene ID. Unlike transcript
version number, it's up to you whether to include gene version number.}
}
\value{
A data frame with at least 2 columns: \code{gene} for gene ID,
\code{transcript} for transcript ID, and optionally \code{gene_name}
for gene names. If \code{other_attrs} has been specified, then those will
also be columns in the data frame returned.
}
\description{
Bioconductor provides Ensembl genome annotation in \code{AnnotationHub}; older
versions of Ensembl annotation can be obtained from packages like
\code{EnsDb.Hsapiens.v86}. This is an alternative to querying Ensembl with
biomart; Ensembl's server seems to be less stable than that of Bioconductor.
However, more information and species are available on Ensembl biomart than
on \code{AnnotationHub}.
}
\examples{
library(EnsDb.Hsapiens.v86)
tr2g_EnsDb(EnsDb.Hsapiens.v86, use_transcript_version = FALSE,
  use_gene_version = FALSE)
}
\seealso{
Other functions to retrieve transcript and gene info: \code{\link{sort_tr2g}},
  \code{\link{tr2g_TxDb}}, \code{\link{tr2g_ensembl}},
  \code{\link{tr2g_fasta}}, \code{\link{tr2g_gff3}},
  \code{\link{tr2g_gtf}}, \code{\link{transcript2gene}}
}
\concept{functions to retrieve transcript and gene info}
