% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr2g.R
\name{tr2g_TxDb}
\alias{tr2g_TxDb}
\title{Get transcript and gene info from TxDb objects}
\usage{
tr2g_TxDb(txdb)
}
\arguments{
\item{txdb}{A \code{\link{TxDb}} object with gene annotation.}
}
\value{
A data frame with 2 columns: \code{gene} for gene ID, \code{transcript}
for transcript ID. For TxDb packages from Bioconductor, gene ID is Entrez ID,
while transcript IDs are Ensembl IDs with version numbers for
\code{TxDb.Hsapiens.UCSC.hg38.knownGene}.
}
\description{
The genome and gene annotations of some species can be conveniently obtained
from Bioconductor packages. This is more convenient than downloading GTF
files from Ensembl and reading it into R. In these packages, the gene
annotation is stored in a \code{\link{TxDb}} object, which has standardized
names for gene IDs, transcript IDs, exon IDs, and so on, which are stored in
the metadata fields in GTF and GFF3 files, which are not standardized.
This function extracts transcript and corresponding gene information from
gene annotation stored in a \code{\link{TxDb}} object.
}
\examples{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
tr2g_TxDb(TxDb.Hsapiens.UCSC.hg38.knownGene)
}
