% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr2g.R
\name{transcript2gene}
\alias{transcript2gene}
\title{Map Transcript ID to Gene ID}
\usage{
transcript2gene(species, kallisto_out_path)
}
\arguments{
\item{species}{A character vector of Latin names of species present in this
scRNA-seq dataset.}

\item{kallisto_out_path}{Path to the \code{kallisto bus} output directory.}
}
\value{
A data frame with two columns: \code{gene} and \code{transcript},
with Ensembl gene and transcript IDs (with version number), in the same order
as in the transcriptome index used in \code{kallisto}.
}
\description{
In \code{kallisto bus}, the Ensembl transcript ID (with version number) is
used, but usually users are interested in number of UMI per gene rather than
transcript. This function queries biomart to get the gene IDs (with version
number) corresponding to the transcripts in the transcriptome assembly.
}
\details{
This function applies only to species that have gene and transcript IDs on
Ensembl.
}
