% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr2g.R
\name{transcript2gene}
\alias{transcript2gene}
\title{Map Ensembl transcript ID to gene ID}
\usage{
transcript2gene(species, fasta_file, kallisto_out_path,
  other_attrs = NULL, ensembl_version = NULL, save = FALSE,
  file_save = "./tr2g_sorted.csv", verbose = TRUE, ...)
}
\arguments{
\item{species}{A character vector of Latin names of species present in this
scRNA-seq dataset. This is used to retrieve Ensembl information from biomart.}

\item{fasta_file}{Character vector of paths to the transcriptome FASTA files
used to build the kallisto index. Exactly one of \code{species} and
\code{fasta_file} can be missing.}

\item{kallisto_out_path}{Path to the \code{kallisto bus} output directory.}

\item{other_attrs}{Character vector. Other attributes to get from Ensembl,
such as gene symbol and position on the genome.}

\item{ensembl_version}{Integer version number of Ensembl (e.g. 94 for the
October 2018 release). This argument defaults to \code{NULL}, which will use
the current release of Ensembl. Use
\code{\link[biomaRt]{listEnsemblArchives}} to see the version number corresponding
to the Ensembl release of a particular date. The version specified here must
match the version of Ensembl where the transcriptome used to build the
kallisto index was downloaded.}

\item{save}{Whether to save the output.}

\item{file_save}{File name of the file to be saved. If the directory in which
the file is to be saved does not exist, then the directory will be created.}

\item{verbose}{Whether to display progress. Defaults to \code{TRUE}.}

\item{...}{Othe arguments to be passed to \code{\link[biomaRt]{useEnsembl}},
such as host and mirror.
Use \code{\link[biomaRt]{listAttributes}} to see which attributes are available.}
}
\value{
A data frame with two columns: \code{gene} and \code{transcript},
with Ensembl gene and transcript IDs (with version number), in the same order
as in the transcriptome index used in \code{kallisto}.
}
\description{
This function is a shortcut to get the correctly sorted data frame with
transcript IDs and the corresponding gene IDs from Ensembl biomart or Ensembl
transcriptome FASTA files. For biomart query, it calls
\code{\link{tr2g_ensembl}} and then \code{\link{sort_tr2g}}. For FASTA files,
it calls \code{\link{tr2g_fasta}} and then \code{\link{sort_tr2g}}. Unlike in
\code{\link{tr2g_ensembl}} and \code{\link{tr2g_fasta}}, multiple species can
be supplied if cells from different species were sequenced together. This
function should only be used if the kallisto inidex was built with
transcriptomes from Ensembl. Also, if querying biomart, please make sure to set
\code{ensembl_version} to match the version where the transcriptomes were
downloaded.
}
\details{
Note that here, the arguments passed to \dots will be passed to
\code{\link[biomaRt]{useEnsembl}} rather than \code{\link[data.table]{fwrite}},
so default settings of \code{\link[data.table]{fwrite}} will be used if
\code{save = TRUE}, which will save a csv file that includes the column names
to \code{file_save}.
}
\examples{
\dontrun{
# Download dataset already in BUS format
library(TENxhgmmBUS)
download_hgmm(".", "hgmm100")
tr2g <- transcript2gene(c("Homo sapiens", "Mus musculus"),
                        kallisto_out_path = "./out_hgmm100")
}

}
\seealso{
Other functions to retrieve transcript and gene info: \code{\link{sort_tr2g}},
  \code{\link{tr2g_ensembl}}, \code{\link{tr2g_fasta}},
  \code{\link{tr2g_gff3}}, \code{\link{tr2g_gtf}}
}
\concept{functions to retrieve transcript and gene info}
