% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{validate_velocity_input}
\alias{validate_velocity_input}
\title{Validate input to get_velocity_files}
\usage{
validate_velocity_input(L, genome, transcriptome, out_path)
}
\arguments{
\item{L}{Length of the biological read. For instance, 10xv1: 98 nt,
10xv2: 98 nt, 10xv3: 91 nt, Drop-seq: 50 nt. If in doubt check read length
in a fastq file for biological reads with the \code{bash} commands:
If the fastq file is gzipped, then do \code{zcat your_file.fastq.gz | head} on
Linux. If on Mac, then \code{zcat < your_file.fastq.gz | head}. Then you will see
lines with nucleotide bases. Copy one of those lines and determine its length
with \code{\link{str_length}} in R or \code{echo -n <the sequence> | wc -c} in
\code{bash}. Which file corresponds to biological reads depends on the particular
technology.}

\item{genome}{Either a \code{\link{BSgenome}} or a \code{\link{XStringSet}}
object of genomic sequences, where the intronic sequences will be extracted
from. Naming convention of chromosomes in the genome must match that in the
annotation, e.g. consistently use 1 for chromosome 1 or consistently use chr1.}

\item{transcriptome}{Either a \code{\link{XStringSet}} or a path to a fasta
file (can be gzipped) of the transcriptome, which contains sequences of
spliced transcripts. This will be concatenated with the intronic sequences
to give one fasta file. The type of transcript ID in the transcriptome must
match that in the gene annotation supplied via argumennt \code{X}. When \code{X} is a
\code{TxDb} object, check the column \code{TXNAME} rather than \code{TXID}. For the
exported function \code{\link{get_velocity_files}}, this argument can be
missing. If it is missing, then the transriptome sequences will be extracted
from the genome given the gene annotation.}

\item{out_path}{Directory to save the outputs written to disk. If this
directory does not exist, then it will be created.}
}
\value{
Nothing. Will throw error if validation fails.
}
\description{
Validate input to get_velocity_files
}
